requirements.txt
flask
flask-sqlalchemy
mysqlclient
pymysql
flask_script
